Stop shipping /usr/sbin/sln
A whole lot has changed since `sln` was introduced into glibc upstream with
this comment:

```
    1998-07-21 07:10  H.J. Lu  <hjl@gnu.org>

            * elf/sln.c: New file.
```
Upstream glibc commit 9d141cae00c957096045d08fe504b4a2be66ab75
```
 Author:     Ulrich Drepper <drepper@redhat.com>
 AuthorDate: Thu Aug 13 19:44:05 1998 +0000
```

Among other things for example, dynamic linking has been in use for the nearly
20 years that have passed and is now quite reliable. Further, many Fedora
systems now use an initramfs (dracut) which is basically a large filesystem of
utilities as a single blob; while it would require rebooting, one can easily
make hardlinks there.

Futher of course there's the increasing use of ostree in Fedora which which
consistently captures *all* of userspace and supports rollback, and additionaly
its presents a "read only" nature to admin tools and other processes that adds
significantly to system resilience.

The failure scenarios that could be recovered via `sln` seem to me to be very
small.

Applying this patch allows dropping a downstream Fedora patch to hardlink `sln` to `ldconfig`,
which was the source of a tricky bug:
https://github.com/projectatomic/rpm-ostree/pull/1173#issuecomment-355014583
that also impacts rpm-ostree.  In practice today SELinux policy is fixed
in updates, but given there's no real reason for this binary to exist
anymore, let's drop it.