Suggest installing minimal localization.
In a discussion with Stephen Gallagher <sgallagh@redhat.com> we
identified that it has been 4 years since we transitioned to
language packs and that at this point we could switch the suggestion
from glibc-all-langpacks to glibc-minimal-langpack which would
prevent an installation from ever accidentally having the larger
installation with all language packs.

This does not yet fix bug 1380069 which is about the removal of
glibc language packs and the way in which the virtual provide is
used.