Deprecate nscd (#1905135)
nscd is planned for a phased removal, with deprecation in Fedora 34:
https://fedoraproject.org/wiki/Changes/DeprecateNSCD

... and removal in Fedora 35:
https://fedoraproject.org/wiki/Changes/RemoveNSCD

systemd-resolved is now enabled by default for DNS caching in Fedora,
and sssd is capable of caching the remaining named services that nscd
handles.  It is therefore time to retire nscd in Fedora and move to more
modern named services caches.